Bugfix: use of deprecated function summarise in dplyr
summarise() must return a value of size 1 since dplyr version 1.1.0. The ifelse()-statement returned a vector- or in case no lab parameters are selected, visits_non_missing was NULL. The now implemented unique()-function with checking if visits_non_missing is NULL, ensures that a vector of length 1 is returned. Now the respective warning disappears.