Two issues were fixed:
Sorting of patient specific values didn't show correct x-values due to unnecessary sort-function in two places.

In some cases, when lines were supposed to be drawn, there was an error due to wrong dimensions, when the x matrix was created. This was caused by the deletion of incomplete cases before removing visits with too many missing values in the tmp dataset, leading to inconsistencies between tmp and raw. Now the order was changed (first, remove visits, then check for complete cases).